Performing sanitization on export
Sort the downloaded file content alphabetically and remove comments from the file
