--- scripts/poe/tests/translations/bad-en.properties
@@ -1,3 +1,0 @@
-one one=one
-two=dos
-three three
