+++ scripts/poe/tests/translations/messages-en.properties
@@ -1,0 +1,3 @@
+Accept\ plain-text\ messages = Accept plain-text messages
+Number\ of\ facilities = {{number}} places
+Number\ of\ form\ types = {{number}} form types
