+++ scripts/poe/tests/translations/messages-sw.properties
@@ -1,0 +1,3 @@
+Accept\ plain-text\ messages = itikia ujumbe wa kawaida
+Number\ of\ facilities = {{nambari}} mahali
+Number\ of\ form\ types = {{nambari}} ya aina fomu
